requirements.txt
torch>=2.2
torchvision>=0.17
torchaudio>=2.2
transformers>=4.45
sentencepiece>=0.2.0
librosa>=0.10.2
soundfile>=0.12.1
numpy>=1.26
scikit-learn>=1.4
matplotlib>=3.8
seaborn>=0.13
timm>=1.0.9
opencv-python>=4.10
einops>=0.7.0
torchmetrics>=1.4
onnxruntime>=1.19
onnxruntime-gpu>=1.19; platform_system == "Windows"
pyyaml>=6.0.1
rich>=13.7
psutil>=5.9
